+++ database/migrations/2020_02_19_094942_remove_unique_cite_on_departures_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class RemoveUniqueCiteOnDeparturesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('departures', function (Blueprint $table) {
+            $table->string('cite')->nullable()->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('departures', function (Blueprint $table) {
+            $table->string('cite')->unique()->nullable()->change();
+        });
+    }
+}